docker/codebase-indexer/requirements.txt
uvicorn
fastapi
GitPython
pydantic
qdrant-client
python-dotenv
requests
pathspec
tree-sitter
tree-sitter-python
tree-sitter-typescript
rich
tqdm
numpy
pandas
torch
transformers
sentence-transformers
watchdog
click
pytest